feat: Enhance login and logout functionality with local storage management
This commit introduces customizations for managing user account information in local storage during login and logout processes. It includes the removal of accountId and userId on logout, and the storage of these values upon successful login, along with pushing a login event to the GTM dataLayer for improved tracking. These changes are part of the Verndale integration to enhance user experience and analytics.